PIM-3832: rename formatter classes and update references
