+++ src/Pim/Bundle/BaseConnectorBundle/Resources/config/converters.yml
@@ -1,0 +1,8 @@
+parameters:
+    pim_base_connector.converter.csv_product.class: Pim\Bundle\BaseConnectorBundle\Reader\File\Converter\CsvProductToStandardConverter
+
+services:
+    pim_base_connector.converter.csv_product:
+        class: %pim_base_connector.converter.csv_product.class%
+        arguments:
+            - '@pim_transform.builder.field_name'
