--- src/Pim/Bundle/BaseConnectorBundle/Resources/config/formaters.yml
@@ -1,8 +1,0 @@
-parameters:
-    pim_base_connector.formater.csv.class: Pim\Bundle\BaseConnectorBundle\Reader\File\Formater\CsvFormater
-
-services:
-    pim_base_connector.formater.csv:
-        class: %pim_base_connector.formater.csv.class%
-        arguments:
-            - '@pim_transform.builder.field_name'
