+++ src/Pim/Bundle/BaseConnectorBundle/Resources/config/processors.yml
@@ -48,7 +48,7 @@ services:
             - '@pim_validator'
             - '@akeneo_storage_utils.doctrine.object_detacher'
             - '@pim_catalog.builder.product'
-            - '@pim_base_connector.formater.csv'
+            - '@pim_base_connector.converter.csv_product'
             - '@pim_catalog.updater.product'
             - %pim_catalog.entity.product.class%
             - 'csv'
